labs/requirements.txt
mysql-connector==2.2.9
python-dotenv==1.0.0
tqdm==4.64.1
